Moved route guards to dedicated source folder
